move toggleLink to lexical/link
