+++ packages/lexical-link/src/index.ts
@@ -19,9 +19,15 @@ import type {
 } from 'lexical';
 
 import {addClassNamesToElement} from '@lexical/utils';
-import {$isElementNode, createCommand, ElementNode} from 'lexical';
-import {Spread} from 'libdefs/globals';
-import invariant from 'shared/invariant';
+import {Spread} from 'globals';
+import {
+  $getSelection,
+  $isElementNode,
+  $setSelection,
+  createCommand,
+  ElementNode,
+} from 'lexical';
+import invariant from 'shared-ts/invariant';
 
 export type SerializedLinkNode = Spread<
   {
@@ -210,3 +216,109 @@ export function $isAutoLinkNode(
 
 export const TOGGLE_LINK_COMMAND: LexicalCommand<string | null> =
   createCommand();
+
+export function toggleLink(url: null | string): void {
+  const selection = $getSelection();
+
+  if (selection !== null) {
+    $setSelection(selection);
+  }
+
+  const sel = $getSelection();
+
+  if (sel !== null) {
+    const nodes = sel.extract();
+
+    if (url === null) {
+      // Remove LinkNodes
+      nodes.forEach((node) => {
+        const parent = node.getParent();
+
+        if ($isLinkNode(parent)) {
+          const children = parent.getChildren();
+
+          for (let i = 0; i < children.length; i++) {
+            parent.insertBefore(children[i]);
+          }
+
+          parent.remove();
+        }
+      });
+    } else {
+      // Add or merge LinkNodes
+      if (nodes.length === 1) {
+        const firstNode = nodes[0];
+
+        // if the first node is a LinkNode or if its
+        // parent is a LinkNode, we update the URL.
+        if ($isLinkNode(firstNode)) {
+          firstNode.setURL(url);
+          return;
+        } else {
+          const parent = firstNode.getParent();
+
+          if ($isLinkNode(parent)) {
+            // set parent to be the current linkNode
+            // so that other nodes in the same parent
+            // aren't handled separately below.
+            parent.setURL(url);
+            return;
+          }
+        }
+      }
+
+      let prevParent = null;
+      let linkNode = null;
+
+      nodes.forEach((node) => {
+        const parent = node.getParent();
+
+        if (
+          parent === linkNode ||
+          parent === null ||
+          ($isElementNode(node) && !node.isInline())
+        ) {
+          return;
+        }
+
+        if ($isLinkNode(parent)) {
+          linkNode = parent;
+          parent.setURL(url);
+          return;
+        }
+
+        if (!parent.is(prevParent)) {
+          prevParent = parent;
+          linkNode = $createLinkNode(url);
+
+          if ($isLinkNode(parent)) {
+            if (node.getPreviousSibling() === null) {
+              parent.insertBefore(linkNode);
+            } else {
+              parent.insertAfter(linkNode);
+            }
+          } else {
+            node.insertBefore(linkNode);
+          }
+        }
+
+        if ($isLinkNode(node)) {
+          if (linkNode !== null) {
+            const children = node.getChildren();
+
+            for (let i = 0; i < children.length; i++) {
+              linkNode.append(children[i]);
+            }
+          }
+
+          node.remove();
+          return;
+        }
+
+        if (linkNode !== null) {
+          linkNode.append(node);
+        }
+      });
+    }
+  }
+}
